cp long id, org post

+++ Backend/Base/Audit/Ent/AuditEnt.cs
@@ -4,12 +4,12 @@ namespace Backend.Base.Audit.Ent
 {
     public class AuditEnt
     {
-        public int Id {  get; set; }
-        public int OrgId { get; set; }
+        public long Id {  get; set; }
+        public long OrgId { get; set; }
         public int Source {  get; set; }
         public int EntityTypeId {  get; set; }
-        public int? EntityId { get; set; }
-        public int UserId { get; set; }
+        public long? EntityId { get; set; }
+        public long UserId { get; set; }
         public DateTime Created {  get; set; }
 	    public string? Crud {  get; set; }
         public string? Details { get; set; }
