+++ Backend/Base/Entity/BaseEntity.cs
@@ -10,8 +10,8 @@ namespace Backend.Base.Entity
 {
     public abstract class BaseEntity : Encode
     {
-        public int Id { get; set; }
-        public int OrgId { get; set; }
+        public long Id { get; set; }
+        public long OrgId { get; set; }
         public string Code { get; set; }
         public string? Description { get; set; }
         public DateTime Updated { get; set; }
