+++ Backend/Base/Role/Ent/UserRoleEnt.cs
@@ -10,7 +10,7 @@ namespace Backend.Base.Role.Ent
 {
     public class UserRoleEnt : BaseEntity
     {
-        public int RoleId { set; get; }
+        public long RoleId { set; get; }
 
         public UserRoleEnt() {}
