+++ Common/DTO/AuditDto.cs
@@ -8,11 +8,11 @@ namespace Common.DTO
 {
     public class AuditDto
     {
-        public int Id { get; set; }
-        public int OrgId { get; set; }
+        public long Id { get; set; }
+        public long OrgId { get; set; }
         public int Source { get; set; }
         public string EntityType { get; set; }
-        public int? EntityId { get; set; }
+        public long? EntityId { get; set; }
         public string? User { get; set; }
         public DateTime Created { get; set; }
         public string Crud { get; set; }
